# lib/api-zod/src/generated/types/assessmentDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Assessment } from './assessment';
import type { Question } from './question';

export type AssessmentDetail = Assessment & {
  questions: Question[];
};