# lib/api-zod/src/generated/types/assessmentGenerateInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssessmentGenerateInputProvider } from './assessmentGenerateInputProvider';

export interface AssessmentGenerateInput {
  /** @minimum 1 */
  lessonId: number;
  provider?: AssessmentGenerateInputProvider;
  /**
     * @minimum 5
     * @maximum 5
     */
  questionCount?: number;
}