# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from './activity';

export interface Dashboard {
  lessonCount: number;
  readyLessonCount: number;
  assessmentCount: number;
  homeworkCount: number;
  providerCount: number;
  recentActivity: Activity[];
}