# lib/api-zod/src/generated/types/exercise.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Exercise {
  id: string;
  prompt: string;
  answer: string;
}