# lib/api-zod/src/generated/types/homeworkGenerateInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HomeworkGenerateInputProvider } from './homeworkGenerateInputProvider';

export interface HomeworkGenerateInput {
  /** @minimum 1 */
  lessonId: number;
  /** @nullable */
  assessmentId?: number | null;
  provider?: HomeworkGenerateInputProvider;
}