# lib/api-zod/src/generated/types/lesson.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LessonStatus } from './lessonStatus';

export interface Lesson {
  id: number;
  title: string;
  course: string;
  /** @nullable */
  level?: string | null;
  status: LessonStatus;
  materialCount: number;
  assessmentCount?: number;
  createdAt: Date;
  updatedAt: Date;
}