# lib/api-zod/src/generated/types/lessonInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LessonInputStatus } from './lessonInputStatus';

export interface LessonInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  course: string;
  /** @nullable */
  level?: string | null;
  status?: LessonInputStatus;
}