# lib/api-zod/src/generated/types/materialInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MaterialInputKind } from './materialInputKind';

export interface MaterialInput {
  /** @minLength 1 */
  name: string;
  kind: MaterialInputKind;
  content: string;
}