# lib/api-zod/src/generated/types/providerSettingsInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProviderSettingsInputProvider } from './providerSettingsInputProvider';

export interface ProviderSettingsInput {
  provider: ProviderSettingsInputProvider;
  /** @nullable */
  model?: string | null;
}